requirements.txt
protobuf~=5.29.6
python-dotenv~=1.2.1
google-genai~=1.65.0